create help_screen.py
Extract HelpScreen and help_text loading from class FlashcardGUI
